requirements.txt
fastapi==0.115.*
uvicorn==0.30.*
httpx==0.27.*
pillow==10.*
numpy==2.*
scikit-learn==1.5.*
gunicorn==22.*
